+++ CHANGELOG.md
@@ -5,13 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+## [0.0.33] - 2020-12-22
+### Changed
+ - Renamed one command: (base) => (site)
+### Added
+ - Added two new commands:
+	(and {operand})
+	(string {operand})
+
 ## [0.0.32] - 2020-12-04
 ### Added
  - Add a new command:
 	(menu-up)
 
 ## [0.0.31] - 2020-07-10
-### Added
+### Changed
  - No changes
 
 ## [0.0.30] - 2020-07-10